+++ rts/motoko-rts-tests/src/gc/incremental/roots.rs
@@ -30,7 +30,7 @@ unsafe fn check_visit_static_roots(heap: &MotokoHeap, root_ids: &[ObjectIdx]) {
         let object = *field;
         if object.tag() != TAG_REGION {
             let array = object.as_array();
-            if array.len() == root_ids.len() as u32 {
+            if array.len() == root_ids.len() {
                 for index in 0..array.len() {
                     let mutbox_value = array.get(index);
                     let mutbox = mutbox_value.as_mutbox();
@@ -51,7 +51,7 @@ unsafe fn check_visit_continuation_table(heap: &MotokoHeap, continuation_ids: &[
         let object = *field;
         if object.tag() != TAG_REGION {
             let array = object.as_array();
-            if array.len() == continuation_ids.len() as u32 {
+            if array.len() == continuation_ids.len() {
                 assert_eq!(context.len(), 0);
                 for index in 0..array.len() {
                     let element = array.get(index);
